Create ModeratableThing
